Split out capabilityRegistry from standard capabilities to avoid loop
