+++ pkg/loop/internal/core/services/capability/registry/capabilities_registry.go
@@ -1,4 +1,4 @@
-package capability
+package registry
 
 import (
 	"context"
